Updates per comments from @dkirkby:
- Fixed message typo in bin/bossquery

- Removed BOSS_CHUNK_SIZE and hardcoded chunk to be 20k lines; this
has some smallish effect on performance, but should be safe value
for any modern-ish laptop or desktop.  Could be possible to detect
free memory and work from that, but that is for later; and could
not hold unzipped file in memory, also for later.

- Changed progress messaging to be ProgressBar messaging; don't know
table length so can't do percentage at this point.

- Added in more detail re: bossquery sort, what, and where options to
bin/scripts.rst